Add flag to enable/disable Go unit tests
The flag only controls the contents of the GitHub action and the
Makefile. The template doesn't provide a Go unit test template and
`make test` will fail out of the box for a component which has Go unit
tests enabled but no tests defined.

See https://github.com/projectsyn/component-keycloak for an example
component which uses Go unit tests.