Allow disabling patch automerge for selected dependencies
We introduce a new cookiecutter argument which allows users to specify
zero or more regex patterns for dependencies whose patch level updates
shouldn't be automerged.

The template expects that the contents of the new argument is a string
which contains zero or more regex patterns separated by semicolons. The
template will use these patterns as the values for the package rule
field `excludePackagePatterns`.