Allow enabling minor automerge based on patterns
This commit introduces a new cookiecutter argument which is expected to
hold zero or more patterns matching dependencies whose minor level
updates should be automerged.

The template expects that the new argument is a string which contains
zero or more regex patterns separated by semicolons. The resulting list
of patterns is used in field `matchPackagePatterns` of a new package
rule that enables automerge for minor updates.

The template doesn't exclude dependencies whose current version is v0.x
from minor automerging if the dependency name matches one of the
provided patterns.