Allow selectively enabling patch dependency automerge for v0.x dependencies
This commit introduces a new cookiecutter argument
`automerge_patch_v0_regexp_allowlist` which can be used to selectively
enable patch automerging for v0.x dependencies.

The template expects that the content of the new argument is a string
which contains zero or more regexp patterns separated by semicolons.

If global automerging of patch updates for v0.x dependencies is
disabled, but `automerge_patch_v0_regexp_allowlist` isn't empty, the
template generates an additional package rule that enables automerge
only for dependencies that are currently v0.x and which match one of the
patterns provided in the argument.

This rule is not required and therefore not generated when
`automerge_patch_v0` is enabled. If you want to disable automerging for
some v0.x patch level updates but automerge patch level updates for v0.x
dependencies by default, you can specify patterns for which patch level
updates shouldn't be automerged in cookiecutter argument
`automerge_patch_regexp_blocklist`.